// packages/sanity/sanity.types.ts
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};

export type BlogCategory = {
  _id: string;
  _type: "blogCategory";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
};

export type About = {
  _id: string;
  _type: "about";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  description?: Content;
  quote?: Content;
  buttonText?: string;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// Source: ./client/queries.ts
// Variable: getTeamMembersQuery
// Query: *[_type == "teamMember"] | order(order asc)
export type GetTeamMembersQueryResult = Array<{
  _id: string;
  _type: "teamMember";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  slug?: Slug;
  intro?: string;
  image?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  twitter?: string;
  website?: string;
  linkedin?: string;
  github?: string;
  instagram?: string;
  twitch?: string;
  order?: number;
}>;
// Variable: getBlogPostsQuery
// Query: *[_type == "blogPost"]{ ..., author-> } | order(publishedAt desc) | order((featured == true) desc)
export type GetBlogPostsQueryResult = Array<{
  _id: string;
  _type: "blogPost";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  slug?: Slug;
  author: {
    _id: string;
    _type: "teamMember";
    _createdAt: string;
    _updatedAt: string;
    _rev: string;
    name?: string;
    slug?: Slug;
    intro?: string;
    image?: {
      asset?: {
        _ref: string;
        _type: "reference";
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
      };
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      _type: "image";
    };
    twitter?: string;
    website?: string;
    linkedin?: string;
    github?: string;
    instagram?: string;
    twitch?: string;
    order?: number;
  } | null;
  featured?: boolean;
  mainImage?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  categories?: Array<{
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    _key: string;
    [internalGroqTypeReferenceTo]?: "blogCategory";
  }>;
  publishedAt?: string;
  body?: Content;
}>;
// Variable: getBlogPostBySlugQuery
// Query: *[_type == "blogPost" && slug.current == $slug][0]{ ..., author-> }
export type GetBlogPostBySlugQueryResult = {
  _id: string;
  _type: "blogPost";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  slug?: Slug;
  author: {
    _id: string;
    _type: "teamMember";
    _createdAt: string;
    _updatedAt: string;
    _rev: string;
    name?: string;
    slug?: Slug;
    intro?: string;
    image?: {
      asset?: {
        _ref: string;
        _type: "reference";
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
      };
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      _type: "image";
    };
    twitter?: string;
    website?: string;
    linkedin?: string;
    github?: string;
    instagram?: string;
    twitch?: string;
    order?: number;
  } | null;
  featured?: boolean;
  mainImage?: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  };
  categories?: Array<{
    _ref: string;
    _type: "reference";
    _weak?: boolean;
    _key: string;
    [internalGroqTypeReferenceTo]?: "blogCategory";
  }>;
  publishedAt?: string;
  body?: Content;
} | null;
// Variable: getBlogPostsSlugsQuery
// Query: *[_type == "blogPost" && defined(slug.current)][].slug.current
export type GetBlogPostsSlugsQueryResult = Array<string | null>;
// Variable: getFusionEventsQuery
// Query: *[_type == "event"] | order(date desc)
export type GetFusionEventsQueryResult = Array<{
  _id: string;
  _type: "event";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  eventType?: "conference" | "meetup";
  date?: string;
  slug?: Slug;
  location?: Location;
  food?: string;
  eventbriteLink?: string;
  topContent?: Array<
    | {
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?:
          | "blockquote"
          | "h1"
          | "h2"
          | "h3"
          | "h4"
          | "h5"
          | "h6"
          | "normal";
        listItem?: "bullet" | "number";
        markDefs?: Array<{
          href?: string;
          _type: "link";
          _key: string;
        }>;
        level?: number;
        _type: "block";
        _key: string;
      }
    | {
        asset?: {
          _ref: string;
          _type: "reference";
          _weak?: boolean;
          [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
        };
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        alt?: string;
        _type: "image";
        _key: string;
      }
  >;
  talks?: Array<
    {
      _key: string;
    } & Talk
  >;
  sponsors?: Array<
    {
      _key: string;
    } & Sponsor
  >;
  bottomContent?: Array<
    | {
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?:
          | "blockquote"
          | "h1"
          | "h2"
          | "h3"
          | "h4"
          | "h5"
          | "h6"
          | "normal";
        listItem?: "bullet" | "number";
        markDefs?: Array<{
          href?: string;
          _type: "link";
          _key: string;
        }>;
        level?: number;
        _type: "block";
        _key: string;
      }
    | {
        asset?: {
          _ref: string;
          _type: "reference";
          _weak?: boolean;
          [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
        };
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        alt?: string;
        _type: "image";
        _key: string;
      }
  >;
  youTubeLink?: string;
}>;
// Variable: getFusionEventBySlugQuery
// Query: *[_type == "event" && slug.current == $slug][0]
export type GetFusionEventBySlugQueryResult = {
  _id: string;
  _type: "event";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  eventType?: "conference" | "meetup";
  date?: string;
  slug?: Slug;
  location?: Location;
  food?: string;
  eventbriteLink?: string;
  topContent?: Array<
    | {
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?:
          | "blockquote"
          | "h1"
          | "h2"
          | "h3"
          | "h4"
          | "h5"
          | "h6"
          | "normal";
        listItem?: "bullet" | "number";
        markDefs?: Array<{
          href?: string;
          _type: "link";
          _key: string;
        }>;
        level?: number;
        _type: "block";
        _key: string;
      }
    | {
        asset?: {
          _ref: string;
          _type: "reference";
          _weak?: boolean;
          [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
        };
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        alt?: string;
        _type: "image";
        _key: string;
      }
  >;
  talks?: Array<
    {
      _key: string;
    } & Talk
  >;
  sponsors?: Array<
    {
      _key: string;
    } & Sponsor
  >;
  bottomContent?: Array<
    | {
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?:
          | "blockquote"
          | "h1"
          | "h2"
          | "h3"
          | "h4"
          | "h5"
          | "h6"
          | "normal";
        listItem?: "bullet" | "number";
        markDefs?: Array<{
          href?: string;
          _type: "link";
          _key: string;
        }>;
        level?: number;
        _type: "block";
        _key: string;
      }
    | {
        asset?: {
          _ref: string;
          _type: "reference";
          _weak?: boolean;
          [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
        };
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        alt?: string;
        _type: "image";
        _key: string;
      }
  >;
  youTubeLink?: string;
} | null;
// Variable: getFusionEventsSlugsQuery
// Query: *[_type == "event" && defined(slug.current)][].slug.current
export type GetFusionEventsSlugsQueryResult = Array<string | null>;
// Variable: getLatestFusionEventQuery
// Query: *[_type == "event"] | order(date desc)[0]
export type GetLatestFusionEventQueryResult = {
  _id: string;
  _type: "event";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  eventType?: "conference" | "meetup";
  date?: string;
  slug?: Slug;
  location?: Location;
  food?: string;
  eventbriteLink?: string;
  topContent?: Array<
    | {
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?:
          | "blockquote"
          | "h1"
          | "h2"
          | "h3"
          | "h4"
          | "h5"
          | "h6"
          | "normal";
        listItem?: "bullet" | "number";
        markDefs?: Array<{
          href?: string;
          _type: "link";
          _key: string;
        }>;
        level?: number;
        _type: "block";
        _key: string;
      }
    | {
        asset?: {
          _ref: string;
          _type: "reference";
          _weak?: boolean;
          [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
        };
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        alt?: string;
        _type: "image";
        _key: string;
      }
  >;
  talks?: Array<
    {
      _key: string;
    } & Talk
  >;
  sponsors?: Array<
    {
      _key: string;
    } & Sponsor
  >;
  bottomContent?: Array<
    | {
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?:
          | "blockquote"
          | "h1"
          | "h2"
          | "h3"
          | "h4"
          | "h5"
          | "h6"
          | "normal";
        listItem?: "bullet" | "number";
        markDefs?: Array<{
          href?: string;
          _type: "link";
          _key: string;
        }>;
        level?: number;
        _type: "block";
        _key: string;
      }
    | {
        asset?: {
          _ref: string;
          _type: "reference";
          _weak?: boolean;
          [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
        };
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        alt?: string;
        _type: "image";
        _key: string;
      }
  >;
  youTubeLink?: string;
} | null;
// Variable: getAboutFusionInfoQuery
// Query: *[_type == "about"][0]
export type GetAboutFusionInfoQueryResult = {
  _id: string;
  _type: "about";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  description?: Content;
  quote?: Content;
  buttonText?: string;
} | null;
// Variable: getCodeOfConductQuery
// Query: *[_type == "codeOfConduct"][0]
export type GetCodeOfConductQueryResult = {
  _id: string;
  _type: "codeOfConduct";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  content?: Content;
  linkText?: string;
} | null;
// Variable: getHomePageContentQuery
// Query: *[_type == "homepageContent"][0]
export type GetHomePageContentQueryResult = {
  _id: string;
  _type: "homepageContent";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  showAlert?: boolean;
  alertContent?: Array<
    | {
        children?: Array<{
          marks?: Array<string>;
          text?: string;
          _type: "span";
          _key: string;
        }>;
        style?:
          | "blockquote"
          | "h1"
          | "h2"
          | "h3"
          | "h4"
          | "h5"
          | "h6"
          | "normal";
        listItem?: "bullet" | "number";
        markDefs?: Array<{
          href?: string;
          _type: "link";
          _key: string;
        }>;
        level?: number;
        _type: "block";
        _key: string;
      }
    | {
        asset?: {
          _ref: string;
          _type: "reference";
          _weak?: boolean;
          [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
        };
        hotspot?: SanityImageHotspot;
        crop?: SanityImageCrop;
        alt?: string;
        _type: "image";
        _key: string;
      }
  >;
} | null;